requirements.txt
alembic==1.16.4
anyio==4.9.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
CacheControl==0.14.3
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
firebase_admin==7.0.0
Flask==3.1.2
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
msgspec==0.19.0
numpy==2.3.1
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.1.1
psycopg==3.2.9
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rich==13.9.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xlsxwriter==3.2.5
